implement warm-restart for checkPodConvergence.py

+++ test_ROM_oligocrystal/rom/checkPodConvergence.py
@@ -42,19 +42,20 @@ logging.info(f'checkPodConvergence.py: Load POD basis in {time.time() - t_local:
 
 for NumFtrs in [1,2,4,8,16,32,64,128,256]:
     for i in range(NumCases):
-        tmpSol = np.zeros([576000,2])
-        predPodCoefs = np.load('../damask/%d/postProc/podCoefs_main_tension_inc%s.npy' % (DamaskIdxs[i], str(PostProcIdxs[i]).zfill(2))) # shape: (5540, 2)
-        # Exclude POD coefs after NumFtrs to check POD convergence
-        predPodCoefs[NumFtrs:,:] = 0
-        # Option: whole domain
-        # tmpSol[:,0] = np.dot(basis_MisesCauchy, predPodCoefs[:,0]) + mean_MisesCauchy
-        # tmpSol[:,1] = np.dot(basis_MisesLnV,    predPodCoefs[:,1]) + mean_MisesLnV
-        # Option: solid domain only
-        tmpSol[SolidIdx,0] = np.dot(basis_MisesCauchy, predPodCoefs[:,0]) + mean_MisesCauchy
-        tmpSol[SolidIdx,1] = np.dot(basis_MisesLnV,    predPodCoefs[:,1]) + mean_MisesLnV
         outFileName = '../damask/%d/postProc/pred_main_tension_inc%s_NumFtrs_%d' % (DamaskIdxs[i], str(PostProcIdxs[i]).zfill(2), NumFtrs)
-        np.save(outFileName, tmpSol)
-        logging.info(f'Processing {i+1:<d}/{NumCases} folders: dumped {outFileName}.npy')
+        if not os.path.exists(outFileName)
+            tmpSol = np.zeros([576000,2])
+            predPodCoefs = np.load('../damask/%d/postProc/podCoefs_main_tension_inc%s.npy' % (DamaskIdxs[i], str(PostProcIdxs[i]).zfill(2))) # shape: (5540, 2)
+            # Exclude POD coefs after NumFtrs to check POD convergence
+            predPodCoefs[NumFtrs:,:] = 0
+            # Option: whole domain
+            # tmpSol[:,0] = np.dot(basis_MisesCauchy, predPodCoefs[:,0]) + mean_MisesCauchy
+            # tmpSol[:,1] = np.dot(basis_MisesLnV,    predPodCoefs[:,1]) + mean_MisesLnV
+            # Option: solid domain only
+            tmpSol[SolidIdx,0] = np.dot(basis_MisesCauchy, predPodCoefs[:,0]) + mean_MisesCauchy
+            tmpSol[SolidIdx,1] = np.dot(basis_MisesLnV,    predPodCoefs[:,1]) + mean_MisesLnV
+            np.save(outFileName, tmpSol)
+            logging.info(f'Processing {i+1:<d}/{NumCases} folders: dumped {outFileName}.npy')
 
 logging.info(f'checkPodConvergence.py: Total elapsed time: {time.time() - t_start} seconds.')
 os.system('htop')
